multiple-images/requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
freeze==1.0.10
Markdown==3.1.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
